docappsystem/requirements.txt
# ================= CORE =================
Django==4.2.23
gunicorn==23.0.0

# ================= DATABASE =================
mysqlclient==2.2.4

# ================= REDIS =================
redis==5.0.1
django-redis==5.4.0

# ================= AWS S3 =================
boto3==1.34.84
botocore==1.34.162
django-storages==1.14.2
s3transfer==0.10.4

# ================= SECURITY =================
cryptography==45.0.6
cffi==1.17.1

# ================= DJANGO UI =================
django-crispy-forms==2.4
crispy-bootstrap5==2025.6
django-bootstrap-v5==1.0.11
django-widget-tweaks==1.5.0

# ================= OPTIONAL FEATURES =================
django-paypal==2.1
pillow==11.3.0

# ================= UTIL =================
python-dotenv==1.0.1
requests==2.32.4

# ================= DJANGO INTERNAL =================
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2

# ================= STATIC FILES =================
whitenoise==6.9.0

# ================= LOW-LEVEL =================
urllib3==2.5.0
charset-normalizer==3.4.3
idna==3.10
certifi==2025.8.3
six==1.17.0
python-dateutil==2.9.0.post0
jmespath==1.0.1
packaging==25.0
typing_extensions==4.14.1
pytz==2025.2
soupsieve==2.7
et_xmlfile==2.0.0
pycparser==2.22
MarkupSafe==3.0.3